Blog_API/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrpt]==1.7.4
python-multipart==0.0.6